better number handling. switched amount to float

+++ database/migrations/2017_03_05_163523_create_transactions_table.php
@@ -15,7 +15,7 @@ class CreateTransactionsTable extends Migration
     {
         Schema::create('transactions', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('amount');
+            $table->float('amount');
             $table->string('type'); // cash, check, credit. debit, etc
             $table->string('location'); // location it was tendered
             $table->date('date');
